// frontend/src/api/models/ModelsGetUserTaskAttemptsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rankode
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ModelsGetUserTaskAttemptsResponse
 */
export interface ModelsGetUserTaskAttemptsResponse {
    /**
     * 
     * @type {string}
     * @memberof ModelsGetUserTaskAttemptsResponse
     */
    code?: string;
    /**
     * 
     * @type {number}
     * @memberof ModelsGetUserTaskAttemptsResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ModelsGetUserTaskAttemptsResponse
     */
    lang?: string;
    /**
     * 
     * @type {number}
     * @memberof ModelsGetUserTaskAttemptsResponse
     */
    memoryUsage?: number;
    /**
     * 
     * @type {number}
     * @memberof ModelsGetUserTaskAttemptsResponse
     */
    runningTime?: number;
    /**
     * 
     * @type {number}
     * @memberof ModelsGetUserTaskAttemptsResponse
     */
    status?: number;
    /**
     * 
     * @type {string}
     * @memberof ModelsGetUserTaskAttemptsResponse
     */
    updatedAt?: string;
}

/**
 * Check if a given object implements the ModelsGetUserTaskAttemptsResponse interface.
 */
export function instanceOfModelsGetUserTaskAttemptsResponse(value: object): value is ModelsGetUserTaskAttemptsResponse {
    return true;
}

export function ModelsGetUserTaskAttemptsResponseFromJSON(json: any): ModelsGetUserTaskAttemptsResponse {
    return ModelsGetUserTaskAttemptsResponseFromJSONTyped(json, false);
}

export function ModelsGetUserTaskAttemptsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelsGetUserTaskAttemptsResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'code': json['code'] == null ? undefined : json['code'],
        'id': json['id'] == null ? undefined : json['id'],
        'lang': json['lang'] == null ? undefined : json['lang'],
        'memoryUsage': json['memory_usage'] == null ? undefined : json['memory_usage'],
        'runningTime': json['running_time'] == null ? undefined : json['running_time'],
        'status': json['status'] == null ? undefined : json['status'],
        'updatedAt': json['updated_at'] == null ? undefined : json['updated_at'],
    };
}

export function ModelsGetUserTaskAttemptsResponseToJSON(json: any): ModelsGetUserTaskAttemptsResponse {
    return ModelsGetUserTaskAttemptsResponseToJSONTyped(json, false);
}

export function ModelsGetUserTaskAttemptsResponseToJSONTyped(value?: ModelsGetUserTaskAttemptsResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'code': value['code'],
        'id': value['id'],
        'lang': value['lang'],
        'memory_usage': value['memoryUsage'],
        'running_time': value['runningTime'],
        'status': value['status'],
        'updated_at': value['updatedAt'],
    };
}